There is one new theorem added to the functor example in examples/HolOmega.
This update also adjusts the Emacs hol-mode file so that it also knows "sml-executable",
which should be the path to the executable for either Moscow ML or Poly/ML as appropriate.
It also updates the developers/releasing-hol script to build a HOL-Omega tar file by the command

./releasing-hol branches/HOL-Omega `which poly`

and which should work also for building the main HOL version.