Closed two holes in the kernel that might have led to soundness bugs. There was no check to see if a new variable or a new constant being created had the base kind. Now these check, and if the kind is not the base kind, an exception is thrown.
This caused serious problems in both kernels in the higher order matching of terms, and in the standard kernel, in how theories are written to disk.  Both of these used the trick of encoding an arbitrary type as a term by just forming a variable of that type.  This meant that both terms and types could be handled in a smooth manner, such as entries on a stack.

However, the larger set of types in HOL-Omega includes types of many other kinds than the base kind, and these cannot be used as the type of a variable, or of any term.  Only proper types, those of the base kind, may be the type of some term.

In the higher order matching of terms, added a more subtle encoding of an arbitrary type <ty> as

<varname> : <ty> (\<tyvarname>:<kind of ty> :<= <rank of ty>. bool)

As long as this is not beta-reduced, it exposes <ty> to the higher order type matching, so its kinds and rank can be matched.

In the standard kernel's method for writing to disk, separated out types into a new shared table just for them.  Then there is no problem with them having non-base kinds.