Implement congruence-based controls for rewriting with the simplifier, to make it easier to rewrite only on the left or right of a top-level relational symbol.  (Strictly, this effect persists recursively down into a term; I don't think this is likely to be an issue, particularly with equality, which I think is the major use-case.)
Now there is an alternative to

  CONV_TAC (LAND_CONV (SIMP_CONV std_ss [...]))

(writing the above has become practically automatic for me, but I
can't help but wince if forced to show it to a new user).

Instead, try

  SIMP_TAC std_ss [..., SimpLHS]

assuming the top-level operator is indeed an equality.  If it's not an
equality, use

  SimpL ``op``

instead.  Will write documentation tomorrow.

Demonstrate use, often in conjunction with Once, in arithmeticScript.