Thm.ETA_CONV -> Drule.ETA_CONV, Thm.Eta -> Drule.RIGHT_ETA
Cursory profiling indicates that an implementation of ETA_CONV as a primitive
inference rule in Thm is about 20 times faster than an implementation as a
derived rule that instantiates ETA_AX, but that the difference in overall HOL
build time is negligible.

Please do let me know if you encounter (performance) problems due to this
change.