Magnus and I added the possibility to throw away all HOL-output (including errors and warnings) inside emacs. This may sound dangerous and useless, since quitedec/interactive turns off all output except warnings and errors. However, it is useful internally sometimes. For example we hide the installation of some auxiliary ml-functions at startup using this method.
Nevertheless, a menu-entry was added to send a region to HOL and ignore all output. For completeness, a command was added as well, that toggles quitedec, sends the region and the toggles it back. Both commands might be useful when opening theories or libraries. The second one usually sends the region and just shows warnings and errors. Depending on the library / theory one is loading there might be a lot of warnings (incomplete pattern matching, free type variables etc). The first command ignores these warnings/errors as well.