Ported revisions 7228:7247 from HOL to HOL-Omega. This is to keep HOL-Omega current with the latest HOL developments.
In addition, added capabilities for coloring both free and type variables, which are initially respectively colored as purple and gold (or brown, for vt100's).  This is different from the type constant color, which is cyan as before.  In addition, type variables and constants have annotations which are revealed by hovering the mouse cursor over them.  These annotations now include kind information, just as the annotations for terms include type information, except that for type constants with the base kind, nothing extra is added since that is the default kind.