Fixed a bug where the parser would not handle errors in the input properly.  If e.g. the input text ended early, the parser would in some situations recover by dropping the part of the input that did not parse, and returning the rest as a valid type, without any error message!
For example,

``:'b (\'a. 'a -> 'a``;

would not throw an exception, despite missing the closing parenthesis, but instead just return the initial portion, as ``:'b``.

The parser strategy has been modified to not recover from such errors, but to permit them to propagate to the place where they are reported.  This involves using a one-token lookahead to decide whether or not the next part of the input is a type operator being applied to the type seen so far.

Thanks to Jeremy Dawson for noticing this bug!