Copied the recent upgrade of type matching to be done the same in the experimental kernel.  The description is duplicated below for convenience.
Revised the higher order type matching algorithm to strengthen it for complex cases of higher order matching.  This increases the power to approximately where it was before the recent removal of divergence.  This does introduce the possibility of divergence again.  Divergence does not occur for the case found earlier by Jeremy Dawson, and it may be precluded entirely, but more analysis of the algorithm is needed to ensure this.

The additional power is provided for cases where a homomorphism is found during the type matching that is not a complete match for a previous homomorphism on the same higher-kind type variable, say 'A.  In such cases, the previous instantiation of 'A is applied as a substitution to the pattern, generating a new pattern.  This new pattern is now matched against the target type as a subproblem of the original type matching problem, where the type instantiations passed to the subproblem are the ones found so far but with 'A substituting for itself (because any instances of 'A now found must be part of the target type, not the pattern type).  This precludes any contrary instantiations of 'A.  If the subproblem succeeds, its instantiations are passed on to the analysis of the rest of the homomorphisms.  If it fails, the algorithm falls back to structural analysis of the pattern type.

In such difficult cases, the prior algorithm just checked to see if the current instantiations found so far were enough to generate the target type from the pattern type, and if so, just discarded the current homomorphism as redundant.  If the target type was not generated, the algorithm fell back to structural analysis of the type.

It should be noted that this algorithm is now far more powerful than the higher order matching algorithm in the standard literature.