Alpha-conversion/naming fix for Cases_on and friends.
The issue was that bound names in the quotation got replaced by bound
names that might appear in the goal.  If, for example, the goal had

  (?x. P x /\ Q x)

somewhere inside it, but you typed

  Cases_on `?y. P y /\ Q y`

you would end up with assumptions ``?x. P x /\ Q x`` in one goal and
``~(?x. P x /\ Q x)`` in the other.