New, powerful, automatic rewrite in optionScript.
Rewrite is

  |- ((if P then SOME x else NONE) = SOME z) = P /\ (x = z)

and all variants there-on.

A number of changes were necessary in other theories.  This change
affects not just srw_ss() but also std_ss.

In finite_map, I created a couple of new Unicode variants for SUBMAP
and FUNION.  I also tidied up some proofs, and the statement of some
goals (indentation, column limits).