Fix of some trouble with the vt100_terminal. I used the 88 color extension. This extension is for example not present in the terminals on Macs. Without this extension, the terminals displayed bound and free variables as blinking(!). I removed the usage of the extension and hope it is working now.
In order to see, how the terminal actually displays annotations and styles, I added a selftest. If you are using a vt100 terminal, it might be interesting to execute this selftest manually several times and play around with extensions like the 88-color one. However, it should hopefully work on old terminals without any extensions now.