Added a comment pointing out that subst can cause accidental variable capture.
This can only occur when the term contains a bound variable that is a
not-yet-generated genvar (which is unlikely to happen unless the user is
malicious) and at least one redex is not a variable. The main issue is that
freshness of genvars is assumed, but not guaranteed by the Term implementation.

Fortunately, all calls to subst in Thm seem to use only variables as redexes,
so there should be no soundness issue. Still, a correct implementation would
seem desirable.