Removed the call to do type-beta reduction after rewriting in GEN_REWRITE_CONV, both in src/bool/Rewrite.sml and in Ho_Rewrite.sml.
Also improved the type comparison functions abconv_ty and abeconv_ty (which is eq_ty) to do a test first of pointer equality on the two types, before reducing them by beta or beta-eta conversion.

This should have  a salutary effect on the timings of the machine_code example, especially the creation of large Hol_datatype structures.