DISCH_ASM_CONV_TAC  and DISCH_ASM_CONSEQ_CONV_TAC moved from quantHeuristics to ConseqConv. Moreover, the implementation has completely changed. Instead of markers from markerLib a new predicate ASM_MARKER defined in ConseqConvTheory is used. This predicate is used extensively to separate all assumptions and make sure, that the original structure of assumptions is preserved.
Moreover, the weight of conversions given to EXT_DEPTH_CONSEQ_CONV has been changed from type int to int option. 
For n > 0, the option SOME n behaves like the old n; NONE behaves like the old argument 0. However, the value SOME 0 provides new functionality.  Conversions with weight NONE are executed till they can not be applied any more, even if the step-count has already exceeded. Weight SOME 0 indicates that these conversions are applied till the step-count is exceeded (and do not count as a step). However, in contrast to NONE as soon as the step-count is exceeded they are not applied any more.