Extended the mllex regexps to allow ] inside a character class (escaped with a backslash).
For example [[\]] matches any square bracket, and [^\]] matches anything except
a closing square bracket. Technically this was possible before, but the new
language is easier. The examples above in the old language would be [[\093] (or
\[|\]) and [\000-\092\094-\128], and these are still acceptable.

This change breaks code where [xyz\093 or [xyz\] is intended to mean [xyz], but
I doubt anyone does that.

I think standard regular expressions also allow ^ and - to be escaped inside a
character class. Mllex will still treat [\^x] as [^x], but that can easily be
changed if desired.