Stop double-reporting of type-check errors in various Q tactics.
The issue is that when parsing becomes part of a tactic application,
the type-checking of the quotation will print out its usual error
message and raise an exception containing the same information.  This
exception is then Raise-d by the proofManagerLib.expand ("e"),
resulting in the same information being printed out twice.  The trick
is turn off the printing of the message during the parse by using the
show_typecheck_errors trace.