Fix an error-reporting bug in the term-parser.
This bug had been concealed by Moscow ML's failure to put the Option
exception at the top-level, but was present in Poly/ML.  The bug was
in the code that checked to see if a possible rule RHS on the top of
the stack was an instance of a list-pattern (with delimiters and
separators).  It did not cope with possible list-patterns that were
too short.  Because this check happens after normal rules have been
looked for, it only manifested when the input was not well-formed.