Made Vimhol's tactic expansion command a little smarter, and made key maps lowercase.
All commands starting with H now start with h instead. The only remaining
capitals are hL, hS, and hR. The motivation is to make common key mappings all
lowercase, to avoid the shift key.

Expand (he) now removes all occurrences of THEN or THEN1 or ) at the beginning
of the purported tactic expression, and removes all THEN or THEN1 or ( at the
end. More such cleverness may be forthcoming.