Adding an ssfrag to a simpset now adds frag rwts to set's dprocs as context.
This has a few knock-on effects, mainly on use of ARITH_ss, which had
hitherto only been getting context from goal assumptions and theorems
passed to the simp-tactics at the top-level.  The ARITH_ss addctxt
function needs to be a bit fussier in order to avoid picking up
unusable junk.

The intention with the change is to get exported rewrites for
relations other than equality to work, as per the change to
churchboolScript (this should be the first of many).  One change this
then required in add_relsimps was to have it check for looping
rewrites (particularly: if reflexivity of the relation is added as a
normal equality rewrite, which is reasonable, this can't be allowed to
put the relation-rewriter into a loop).