Fix hypothesis-free selftest, by putting right new_constant in scope.
It turns out this test wasn't testing quite what I thought it was because the
call to new_constant was the kernel's and not boolLib's.  The kernel's call
doesn't affect the parser, and so the subsequent parsing in the call to 
new_axiom created foo as a variable, not the constant.