Bug: bounds on rewrites should be shared between all derived rewrites.
In a situation like

  th = |- (p = x) /\ (q = y)

as the input rewrite, you might want separate bounds (i.e., both
conjuncts of Once th get to apply once each), but with

  |- (s1 = s2) <=> (!x. x IN s1 <=> x IN s2)

you surely do not want the two derived theorems

  |- (s1 = s2) <=> (!x. x IN s1 <=> x IN s2)
  |- (s2 = s1) <=> (!x. x IN s1 <=> x IN s2)

to both get a chance to apply.  Arguably, in the latter case, the
simplifier should be smart enough to see that the two terms are
instances of each other and not generate two theorems, but the test is
not as simple as just "are both sides variables", because one or both
of the variables may be bound in a goal's assumptions, and then both
rewrites are useful.