Revised RES_CANON again to make it fully backwards compatible.  Now when the argument is a theorem which contains no higher kind or higher rank type variables, universal types, type abstractions, term-type combinations or abstractions of a term by a type variable, that is, within the normal HOL4 higher order logic, the result is a list of theorems within that logic as well, exactly the same as RES_CANON in HOL4.  This restores full backwards compatibility with HOL4.
For other theorem arguments outside the HOL4 higher order logic, RES_CANON will add universal quantifications of type variables which were free in the original theorem, which is the natural and right thing to do.  If the user desires this behavior for a theorem in the HOL4 logic, just use (RES_CANON o TY_SPEC_ALL).

Also added type version of many conversions in Conv.sml, and added some awareness of the new HOL-Omega logic forms to the MESON_TAC library.