Revision of the implementation of type abbreviations; these are intended, when printing types, to condense the output and increase its readability.
In general, there may be more than one way to interpret a type as a type abbreviation with arguments.  Because of the higher order matching of type arguments, sometimes the representation could actually be larger and more complicated using a particular type abbreviation than without, because the arguments could be complex forms that combine with the abbreviation and reduce under eta-beta reduction to form a simple type.  In such cases, the type abbreviation is not used, as it does not help.  This helps to reduce spurious and unhelpful instances of type abbreviations when printing types.  Otherwise, the type

``:\'b. ('a, 'b) state``

in examples/HolOmega/monadScript.sml may print as

``:((\'b. 'b # 'a) o 'a fun)``

This means, however, that some type abbreviations such as

val _ = type_abbrev ("Kleisli", Type `: \'A 'M 'a 'b. ('a, 'b 'M) 'A`) ;

are unlikely to be used in printing types, as the size of ('A,'M,'a,'b)Kleisli is larger than ('a, 'b 'M) 'A.