Fix handling of special 'U' constant, retaining better backwards compatibility.
In particular, there is now a special form 'univ' that behaves as 'U' did.
Further, with Unicode on, the character U+1D54C is used in the same way.
This character may not be in many fonts, so add another trace to turn its
use off.  If you can see it, it's a nice 'double-struck' U.  I would have
gone for a script-style U, but it leant over too much, and so overlapped
the following left-paren character.