Implement a new (automatic) simplification for terms like {(x,y)|F}.
Such terms rewrite to {}, but with the underlying paired abstractions
getting in the way, it's impossible to get a normal rewrite to do the
right thing.  The implemented conversion also tries to handle the
situation where the predicate is T, but this isn't sure to rewrite to
UNIV because pattern-forms may preclude it.  For example,
  { x + 1 | T }
with x a natural, is not the universal set.

The simpset fragment has name GSPEC_SIMP if you need to remove it.
The distribution builds at selftest level 1.  Selftest level 2 will be
tested overnight.