Update use of exec to reflect its proper practice (name of executable in args).
When you call execv (the Unix system call), you pass the thing to run as well
as an argv.  Tradition has that the first thing in argv is the name of the
thing being executed (in shell scripts, this is bash, for example).  This change
makes our use of exec match this.