Fix another Unicode-bounce bug.
The real problem in the previous bug was that the Unicode flag was being
bounced as a term was printed for inclusion in the generated theory file.
The previous fix merely removed the bouncing.  This fix deals with the
code around the bouncing properly, and simplifies the handling of Unicode
overloads.