Provide very simple rule-induction tactic under Induct_on.
If you do Induct_on `myindrel arg1 arg2`, then it looks up myindrel in its
little database, and attempts a HO_MATCH_MP_TAC on the theorem it finds there
(if there is one).  The intention is to make this a deal more sophisticated
but it now saves users from having to remember induction theorem names.