Adjust the way multiple matching abbreviations are discriminated amongst.
In particular, it used to be that if an abbreviation required a smaller
type instantiation, it would be preferred.  Now, something that requires
a smaller *term* instantiation gets the nod.  Type instantiation size is
the second factor considered, and then recentness.  

Also add raw_print_map entry-point to Overload module to add debugging.