Get ARITH_ss to avoid doing some unnecessary work.
In particular, there is no point in ever getting it to attempt a
goal whose conclusion is of the form p /\ q.  This is because it
will have already attempted p and q independently, and doing them
together can't make the goal any more solvable.  One wrinkle though:
the simpset tries to prove them equal to true, and to false.  So,
if the term to be simplified is p /\ q, it's still worth trying to
prove this is false.