Fix for the bug in Hol_datatype to do with nested types and itself. Also includes another selftest that discriminates between the current fix and a previous non-fix I tried.
The problem, I believe, was in how the actual new types being defined were filtered out of the domains of functions asserted to exist by the recursion theorem.
The old way was to exclude types with an argument appearing somewhere else in the set of candidate types.
The new way is to simply exclude types with non-variable arguments.
I'm not sure this is right, but it works.

The selftests still fail, but for an entirely different reason: the second test tries to redefine the types t1 and t2 from the first test.
Using distinct names would fix this, but I would rather have the ability to redefine a name.