Fix for TypeNet bug (and new regression test in src/bool).
The issue wasn't really anything to do with Hol_datatype.  It was having
problems because of the way it inserts data about new types into the
TypeBase after a type has been defined.  The TypeNet structure is used by
the type-abbreviation facility too, and so the 'same' bug can be
demonstrated using type_abbrev.  The fix to TypeNet was to get it to
explicitly record the arity of the type operators in the nodes of its
internal trie.