Set up ML context in pattern-matching selftest by open-ing bossLib.
This is needed because Moscow ML assembles the object files based on
what it sees in the source file, rather than just working off a
heap as Poly/ML does.  (Poly/ML gets to work off the 'bossLib' heap
because of the fact that src/tfl/src/test appears later in the build-
sequence.