Polish the user experience in Poly's smart-configure.
In particular, if the overrides file exists (and provides sensible values),
don't tell the user to write it when the automatic setting detection fails.
In fact, if the overrides file exists and gives a value for something,
don't even attempt to determine that value automatically.

One slight change though is that you can now only override the poly
and polymllibdir values.

Thanks to Tjark for the report.