Allow Moscow ML build to happen using (as yet unreleased) v2.10.
Basic strategy is to make adoption of basis 2002 even more
thorough.  In particular, the Timer structure is now as per
the 2002 Basis.

It is possible to detect that one is working with Moscow ML 2.10 or
later by checking the Holmakefile variable $(HAVE_BASIS2002), which is
set (to value "1") when in 2.10 or later.  It is unset elsewhere.  See
an example of its use in help/src-sml/Holmakefile.