Correct a minor problem in remembering the ssfrags making up a simpset.
In particular, the old code wouldn't remember that a fragment had been
added (it wouldn't add it at all) if it had the same name as one
already in the set.  This was done with the op_insert same_frag.

Now the code just conses the fragment onto the list.  It might be
better to merge the incoming fragment with one in the set of the same
name, but I don't know that there's really a problem with having
multiples of the same name.

The problem particularly arises when successive calls to
export_rewrites in a theory cause multiple fragments to be added to
the srw_ss() all with the same name (the name of the current theory).