Fix a bug in the handling of weakened 'subset' theorems.
A subset theorem is one of the form |- R' x y ==> R x y where R' is a
relation that might appear in an input theorem, and R is the relation
that one wishes to rewrite with respect to.  When the weakened rewriter
sees a bit of context of the form R' x y, it applies modus ponens to
derive useful information for its R-rewriting.