Handle compilation error for stringSimps; change ssfrag internals.
Clearly, nothing used that file because it was only Moscow ML
compilation that complained about it.  Anyway, I've added a
destructor function to the simpLib API for getting the rewrites out
of a simpset fragment.  I also changed the internal representation
so that SSFRAG is no longer the constructor.  Instead SSFRAG_CON is,
and I will now be able to add additional fields to the record that
it takes as an argument.