BIG refactor: mainly to allow stdknl to use SharingTables technology.
SharingTables implements hash-consing for types as they are written
out to disk.  This can make a big difference to the theory file size,
and thus to theory-loading times.

The big visible change is that Theory, TheoryPP and Definition are no
longer really part of the kernel, and are found in src/postkernel
instead.  They don't need to know about secret implementation details
in terms, types or theorems.  They do need to have access to

  Term.{write,read}_raw

and primitive definition principles (which are in Thm).

Another API change is that Tags no longer use user-visible string refs
to record use of axioms.  Instead, the refs are hidden behind a new
Nonce type.  This in turn means that the Tag.axioms_of function can
move to be visible to all users after the kernel.  (If we kept the
axioms as string refs, users could alter the strings being pointed
to.)