Significant change to pretty-printing; affects user-printing API.
The change allows user-printers to alter the main printer's notion of
what is a "seen free variable", and what is bound.  The latter causes
the colouring.  The first notion is important when show_types is on,
as types are only given for the first occurrence of a free variable.

This allows the monadsyntax to do the right thing with the colouring
of variables.