Add Opentheory module with read_article function
This is an alpha-quality implementation and contains many (unknown) bugs.

Comments on design decisions welcome.

read_article :
  instream ->
  {thms   : thm net,
   types  : (string, hol_type list -> hol_type) dict,
   consts : (string, hol_type -> term) dict} ->
  thm list
takes an article and a record of theorems, types and constants the
article might require, processes it, and returns the exported theorems.

opentheory can tell you what thms types and consts a package would need,
and read_article should raise informative errors if you don't provide
something.

Also include a selftest that processes the article for bool-def-1.0, a
package that appears to come with opentheory in the "gilith" repository.
But I don't yet test that the generated theorems match what the article
is supposed to produce, because that probably first requires
implementing a parser for the output of opentheory info --summary.
(Don't want to list them by hand...)