Complete rewrite of Opentheory reader
It now successfully reads the package bool-1.0 in the gilith repository,
and that is the selftest (only runs if you have opentheory installed).

The selftest file is also an example of the work you might need to do to
read an opentheory article that defines constants that are already
defined in HOL.

When the article is only going to define new things, the work would
likely be less, although some helper functions (TODO) in the Opentheory
module would make it easier.