Example theories GilithBool and GilithUnit
These theories export each export a single theorem that is a conjunction
of all the exports of the articles for bool-1.0 and unit-1.0 in the
gilith repository. They use HOL's already-defined constants. The
GilithUnit example shows one way to use the exports from another
opentheory theory to satisfy the requirements of the current one.