Extend Redblackmap with update and findKey operations
update allows you to set the value for a key to some function of its old
value without traversing the tree twice. findKey returns the inserted
key along with its value, which may be useful if the lookup key is
different but still EQUAL as a key.

Strictly speaking these are not extensions to the original code, since I
have taken the obvious code reuse opportunities. So now find and insert
might waste some cycles constructing and deconstructing options and
closures.