This is a major upgrade to the HOL-Omega logic and system.
The major change is deep but important: whereas ranks were previously an attribute of ki
nds, they are now made an attribute of kinds, specifically of both the base kind and kin
d variables.  This enables the kind of a type operator to have a specific rank for the k
ind of its domain.  This rank can then be used to judge the appropriateness of a given t
ype argument, to see if it is of the expected rank or lower.

This is a new restraint on the application of type operators to type arguments, and has
required a considerable expansion and reworking of the whole type/kind/rank inference sy
stem, as well as other portions of the code.  The benefit is that the rank of type expre
ssions may be now much more accurately estimated than before, and that applications of t
ype operators to arguments which do not respect rank are now detected immediately and re
jected.  In the previous system, such applications were allowed, but had no useful meani
ng according to the official semantics, and in practical terms, could not be beta-reduce
d.

A message will be posted to the Hol-info mailing list soon with more details.