Theory of complex numbers.
E-mailed to me by Liming Li, based on a HOL Light development.
May need to revisit decision to use 'i' as the name for the constant (HOL
Light uses 'ii').  John also has complex as a whole new type; this
development just uses a real#real type abbreviation.  Of course, it's good
to explore the possibilities.