Rework lambda-term recursion principle proof once more.
It's simpler and more polished, mostly following the IJCAR'06 paper by
Urban and Berghofer, though slightly simpler even than that.  I
believe I have one more useful change in me for this development, and
then I'll attempt to do the whole thing again, generically in
genericTerms.

The change to term_posnScript comes about because it relied slightly
too much on the exact form of the principle.  (Previously tm_recursion
was just recursion and not iteration; putting iteration in from the
outset seems harmless.)