Fix (and test) for bug in quotient package: it was parsing with global grammar.
It's OK to parse with the global grammar when in a script file as one
knows then what context one is parsing strings within.  But in library
code, such as quotient.sml, it is dangerous to rely on the behaviour
of the global grammar because users may have changed it to the
library's disadvantage.  Better to do all of one's internal parsing
with a known grammar.